moved checkJavaInstallation to common.ts and fixed lint errors
