+++ src/commands/android/utils/common.ts
@@ -12,6 +12,7 @@ import {ABI, AVAILABLE_OPTIONS, AVAILABLE_SUBCOMMANDS, DEFAULT_CHROME_VERSIONS, 
 import {Platform, SdkBinary} from '../interfaces';
 import Logger from '../../../logger';
 import untildify from 'untildify';
+import {execSync} from 'child_process';
 
 export const getAllAvailableOptions = () => {
   const mainOptions = Object.keys(AVAILABLE_OPTIONS);
@@ -155,74 +156,6 @@ export const downloadFirefoxAndroid = async (version: string) => {
   return await downloadWithProgressBar(apkDownloadUrl, tempdir);
 };
 
-export const showHelp = (unknownOptions: string[], unknownSubcommand?: string) => {
-  if (unknownSubcommand) {
-    Logger.log(colors.red(`unknown subcommand passed: ${unknownSubcommand}\n`));
-  } else if (unknownOptions.length) {
-    Logger.log(colors.red(`unknown option(s) passed: ${unknownOptions.join(', ')}\n`));
-  }
-
-  Logger.log(`Usage: ${colors.cyan('npx @nightwatch/mobile-helper android [options] [subcommand] [subcommand-options]')}`);
-  Logger.log('  Verify if all the requirements are met to run tests on an Android device/emulator.\n');
-
-  Logger.log(`${colors.yellow('Options:')}`);
-
-  const switches = Object.keys(AVAILABLE_OPTIONS).reduce((acc: {[T: string]: string}, key) => {
-    acc[key] = [key].concat(AVAILABLE_OPTIONS[key].alias || [])
-      .map(function(sw) {
-        return (sw.length > 1 ? '--' : '-') + sw;
-      })
-      .join(', ');
-
-    return acc;
-  }, {});
-
-  const longest = (xs: string[]) => Math.max.apply(null, xs.map(x => x.length));
-
-  const switchlen = longest(Object.keys(switches).map(function(s) {
-    return switches[s] || '';
-  }));
-
-  const desclen = longest(Object.keys(AVAILABLE_OPTIONS).map((option) => {
-    return AVAILABLE_OPTIONS[option].description;
-  }));
-
-  Object.keys(AVAILABLE_OPTIONS).forEach(key => {
-    const kswitch = switches[key];
-    let desc = AVAILABLE_OPTIONS[key].description;
-    const spadding = new Array(Math.max(switchlen - kswitch.length + 3, 0)).join('.');
-    const dpadding = new Array(Math.max(desclen - desc.length + 1, 0)).join(' ');
-
-    if (dpadding.length > 0) {
-      desc += dpadding;
-    }
-
-    const prelude = '  ' + (kswitch) + ' ' + colors.grey(spadding);
-
-    Logger.log(prelude + ' ' + colors.grey(desc));
-  });
-
-  Logger.log(`\n${colors.yellow('Subcommands and Subcommand-Options:')}`);
-
-  Object.keys(AVAILABLE_SUBCOMMANDS).forEach(subcommand => {
-    const subcmd = AVAILABLE_SUBCOMMANDS[subcommand];
-    const subcmdOptions = subcmd.options.map(option => `[--${option.name}]`).join(' ');
-
-    Logger.log(`  ${colors.cyan(subcommand)} ${subcmdOptions}`);
-    Logger.log(`  ${colors.grey(subcmd.description)}`);
-
-    if (subcmd.options && subcmd.options.length > 0) {
-      const optionLongest = longest(subcmd.options.map(option => `--${option.name}`));
-      subcmd.options.forEach(option => {
-        const optionStr = `--${option.name}`;
-        const optionPadding = new Array(Math.max(optionLongest - optionStr.length + 3, 0)).join('.');
-        Logger.log(`    ${optionStr} ${colors.grey(optionPadding)} ${colors.grey(option.description)}`);
-      });
-    }
-    Logger.log();
-  });
-}
-
 export const getSdkRootFromEnv = (androidHomeInGlobalEnv: boolean, rootDir: string): string => {
   Logger.log('Checking the value of ANDROID_HOME environment variable...');
 
@@ -256,4 +189,22 @@ export const getSdkRootFromEnv = (androidHomeInGlobalEnv: boolean, rootDir: stri
   }
 
   return '';
-}
+};
+
+export const checkJavaInstallation = (rootDir: string): boolean => {
+  try {
+    execSync('java -version', {
+      stdio: 'pipe',
+      cwd: rootDir
+    });
+
+    return true;
+  } catch {
+    Logger.log(`${colors.red('Error:')} Java Development Kit v9 or above is required to work with Android SDKs. Download from here:`);
+    Logger.log(colors.cyan('  https://www.oracle.com/java/technologies/downloads/'), '\n');
+
+    Logger.log(`Make sure Java is installed by running ${colors.green('java -version')} command and then re-run this tool.\n`);
+
+    return false;
+  }
+};
